Refactor UI and separate utils to a new file
